# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: NeuroGrid API
  description: |
    NeuroGrid is a distributed AI computing platform that enables secure and efficient deployment of AI models across a decentralized network.
    
    ## Features
    - Text generation with LLaMA models
    - Image generation with Stable Diffusion
    - Audio transcription with Whisper
    - Real-time task monitoring
    - Secure authentication and authorization
    
    ## Base URLs
    - Production: `https://api.neurogrid.io`
    - Development: `https://api-dev.neurogrid.io`
    
  version: 1.0.0
  contact:
    name: NeuroGrid Support
    url: https://neurogrid.io/support
    email: support@neurogrid.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.neurogrid.io/api/v1
    description: Production server
  - url: https://api-dev.neurogrid.io/api/v1
    description: Development server
  - url: http://localhost:3001/api/v1
    description: Local development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: securepassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh JWT token
      description: Get a new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  example: refresh_token_here
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  # Nodes
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            token:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            refreshToken:
              type: string
              example: refresh_token_here
            expiresIn:
              type: integer
              example: 3600
            user:
              type: object
              properties:
                id:
                  type: string
                  example: user123
                email:
                  type: string
                  example: user@example.com
                role:
                  type: string
                  example: user
                permissions:
                  type: array
                  items:
                    type: string
                  example: [read, write]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_REQUEST
            message:
              type: string
              example: Human readable error message
            type:
              type: string
              example: ValidationError
            details:
              type: object
              description: Additional error details
...